test: update type

+++ controlplane/frn-core/src/compute/instance.rs
@@ -72,8 +72,8 @@ pub struct InstanceCreateRequest {
     /// The number of cores per socket.
     pub cores: u8,
 
-    /// The disk size.
-    pub disk_size: u32,
+    /// The disk size in bytes.
+    pub disk_size: u64,
 
     /// The disk image to create the instance from.
     pub disk_image: String,
